From `eval_property_signatures.py`, we get the following improvements:
* Concrete value signature time: 0.62 -> 0.34 ms per value (1.8x speedup)
* Lambda signature time: 18.29 -> 1.21 ms per value (15x speedup)
* Signature length: 2167 -> 917 (2.36x reduction)
* Concrete signature collisions: 68 -> 44
* “Avoidable” concrete signature collisions: 27 -> 3

I can now train with approximately 1.6 sec / it, on 8 GPU with grad_acc=4 (so 4 synthesis searches per "iteration").

Summary of changes

We run lambdas fewer times:
* Instead of having 32 canonical ints and 32^2=1024 canonical pairs of ints, I've handwritten 16 canonical ints and 16 canonical pairs of ints.
* Instead of running a lambda on all canonical input lists of the appropriate arity for all examples, we only use each input list once and do a round-robin through examples.

Signatures are shorter:
* Because lambdas are run on all canonical input lists (of the appropriate arity), not ignoring any inputs anymore (elaborated below), it is unnecessary to spend signature space describing the inputs themselves (they are constant given the arity).
* I/O examples may have 3 inputs but lambdas have at most 2 inputs. We used to pad all input lists to 3 inputs, but now we pad to 2 for lambdas.
* I removed some properties, targeting those that contributed a lot to the overall signature length but were the least useful in distinguishing different values. This process was informed by stats printed by running `eval_property_signatures.py` and `property_signatures.py`. (I also added a few properties that helped distinguish values without adding much to the length.)

Signatures are faster to process:
* We used to reduce across examples to produce `(frac_applicable, all_true, all_false, frac_true)`. I removed `all_true` and `all_false` since they are redundant with `frac_true` (especially after quantization where 0 and 1 get their own buckets). This means signatures have half as much data
* In the model, instead of processing and expanding (embedding) signatures on CPU with explicit for loops, and then moving the result to GPU, we move signatures to GPU first and then process/embed them using torch operations.

Other changes:
* We used to ignore input lists that caused a lambda to fail, because we were searching for the correct types to give. but we have since realized that all lambda inputs are ints. now, when a lambda fails (and returns None) for some inputs, we use this info in the property signature. i've added a property that says whether an object (in this case, the result of the lambda) is None, and the rest of the properties are N/A (-1 in the code, used to be None). This in turn affects the "fraction applicable" part of the property signature after reducing across examples.
* A Value that is effectively constant (does not use any input variables) has only 1 "example" in it, e.g., `value.values == [25]`. But it's possible that another Value uses input variables and comes up with the same constant, e.g., `value.values == [25, 25, 25]` for 3 examples. The search treats these values as different, as their repr is different. I've changed the repr to be the same, so these values are treated as equal.
* Change eval_property_signatures.py distance analysis to make it faster (don't process all pairs)
* Added some profiling code
* Added more statistics for the analysis script (which I'll put in a separate PR)